Phew. Everything working. Clean as a whistle.
The worst kind of bug is the one which you fix and still have no
idea what went wrong in the first place. This was one of the
instance. Look out the commented code in Peer_Clientc. to get a
glimpse of what was bothering. Maybe one day I will get a sudden
realisation why I had to spend 3 hours fixing this "bug". Part of
it was my stupidity ofcourse, the rest can be summarised by saying
SHIT HAPPENS.