Distributing through cord instead of the server.
Using a signal from user Ctrl-Z to send the START message from
the main process to the nodes. This is done because, you never know
when the file distirbution completes as the number of files and nodes
can be really big. Therefore, for now a manual signal is being used to
make the nodes start accepting requests.

This should be changed later to a more sophisticated approach.
For Assignment 7, the topology initiation should be changed and the
rest will hopefully remain the same.

Cleaning of stdout should be done.